// src/generated/revalidate-itinerary.ts
export interface components {
    schemas: {
// ...
        BaggageAllowanceType: {
            /**
             * @description The first line of the description of the baggage allowance type.
             * @example UP TO 50 POUNDS/23 KILOGRAMS
             */
            description1?: string;
            /**
             * @description Baggage text information for allowance. The second line of the description of the baggage allowance type.
             * @example UP TO 62 LINEAR INCHES/158 LINEAR CENTIMETERS
             */
            description2?: string;
            /**
             * @description The unique ID represents the baggage allowance type.
             * @example 1
             */
            id: number;
            /**
             * @description The total amount of baggage pieces allowed on the journey.
             * @example 1
             */
            pieceCount?: number;
            /**
             * @description Allowed baggage weight unit indicator. Possible values are lbs, kg.
             * @example lbs
             */
            unit?: string;
            /**
             * @description The total allowed weight of the baggage.
             * @example 20
             */
            weight?: number;
        };
// ...
        BaggageInformationType: {
            /**
             * @description The code used to identify the airline in charge of the baggage.
             * @example AA
             */
            airlineCode: string;
            /** @description Reference to Baggage Allowance ID. */
            allowance?: components["schemas"]["Allowance"];
            /** @description Reference to Baggage Charge ID. */
            charge?: components["schemas"]["Charge"];
            /**
             * @description Contains baggage information, e.g. allowance, charges, embargo, etc. A - Check-in Allowance, C - Day of Check-in Charges, etc.
             * @example A
             */
            provisionType: string;
            /** @description Index of the segment covered by this baggage information relative to the beginning of the itinerary. Starts from 0. */
            segments: components["schemas"]["Segment"][];
        };
// ...
        GroupedItineraryResponse: {
            /**
             * @description Grouped Itinerary Response.
             * @example {
             *       "Version": "V5"
             *     }
             */
            groupedItineraryResponse: {
                /** @description Baggage Allowance information. */
                baggageAllowanceDescs?: components["schemas"]["BaggageAllowanceType"][];
                /** @description Baggage Allowance information. */
                baggageChargeDescs?: components["schemas"]["BaggageChargeType"][];
                /** @description Features Table (166) information (ancillaries attached to brand). */
                brandFeatureDescs?: components["schemas"]["BrandFeatureType"][];
                /** @description Fare calculation information. */
                fareComponentDescs?: components["schemas"]["FareComponentType"][];
                /** @description Lists all Flight Amenities of different types that are available on any of the returned flights. Amenities are indexed to easily associate them with flight components. They are later referenced from other elements. */
                flightAmenities?: components["schemas"]["FlightAmenitiesListType"];
                /** @description Itinerary group. */
                itineraryGroups?: components["schemas"]["ItineraryGroupType"][];
                /** @description Leg ID description. */
                legDescs?: components["schemas"]["LegType"][];
                /** @description Server Message. */
                messages: components["schemas"]["MessageType"][];
                /** @description OB Fee information. */
                obFeeDescs?: components["schemas"]["OBFeeType"][];
                /** @description Summary of taxes calculated for original itinerary during Exchange transaction. */
                originalItineraryDesc?: components["schemas"]["OriginalItineraryDescType"];
                passengerDescs?: components["schemas"]["PassengerDataType"][];
                /** @description List of Price Classes referenced from the NDC offers. */
                priceClassDescriptions?: components["schemas"]["PriceClassDescriptionsType"][];
                /** @description Schedule information. */
                scheduleDescs?: components["schemas"]["ScheduleDescType"][];
                /** @description Schedule Global Message. */
                scheduleMessages?: components["schemas"]["ScheduleMessageType"][];
                /** @description Response statistics. */
                statistics?: components["schemas"]["StatisticsType"];
                /** @description Tax information. */
                taxDescs?: components["schemas"]["TaxType"][];
                /** @description Tax summary information. */
                taxSummaryDescs?: components["schemas"]["TaxSummaryType"][];
                /** @description Universal Product Attribute category information. */
                upaCategoryDescs?: components["schemas"]["UPACategoryDescType"][];
                /** @description Universal Product Attribute information. */
                upaDescs?: components["schemas"]["UPADescType"][];
                /** @description Universal Product Attribute photo information. */
                upaPhotoDescs?: components["schemas"]["UPAPhotoDescType"][];
                /** @description Lists Universal Product Attribute seat characteristics. */
                upaSeatCharacteristicDescs?: components["schemas"]["UPASeatCharacteristicDescType"][];
                /** @description Universal Product Attribute tour information. */
                upaTourDescs?: components["schemas"]["UPAMediaDescType"][];
                /** @description Universal Product Attribute video information. */
                upaVideoDescs?: components["schemas"]["UPAMediaDescType"][];
                /** @description Universal Ticketing Attribute information. */
                utaDescs?: components["schemas"]["UniversalTicketingAttributeType"][];
                /** @description Validating Carrier information. */
                validatingCarrierDescs?: components["schemas"]["ValidatingCarrierType"][];
                /**
                 * @description Required response version of the service.
                 * @example V5
                 */
                version: string;
            };
        };
// ...
        LegType: {
            /**
             * @description The elapsed time at the LegDesc level in GIR format. This information allows customers to display travel time for NDC and ATPCO Offers.
             * @example 471
             */
            elapsedTime?: number;
            /**
             * @description Leg ID information.
             * @example 2
             */
            id?: number;
            /** @description Schedule reference, plus departure day offset. */
            schedules: components["schemas"]["ScheduleType"][];
        };
// ...
        "OTA_AirLowFareSearchRQ.OriginDestinationInformation.TPA_Extensions": {
// ...
            Flight?: {
                Airline: components["schemas"]["OTA_AirLowFareSearchRQ.OriginDestinationInformation.TPA_Extensions.Flight.Airline"];
                /**
                 * @description Arrival date and time.
                 * @example 2024-08-01T06:30:00
                 */
                ArrivalDateTime: string;
                /**
                 * @description This attribute within the flight details information is required to enable the verification path, only for the Revalidate Itinerary Functionality.
                 * @example Y
                 */
                ClassOfService?: string;
                /**
                 * @description Departure date and time.
                 * @example 2024-08-01 09:50:00
                 */
                DepartureDateTime: string;
                DestinationLocation: components["schemas"]["OTA_AirLowFareSearchRQ.OriginDestinationInformation.TPA_Extensions.Flight.DestinationLocation"];
                /**
                 * Format: int16
                 * @description Number of the flight.
                 * @example 101
                 */
                Number: number;
                /** @description Departure location code. */
                OriginLocation: components["schemas"]["OTA_AirLowFareSearchRQ.OriginDestinationInformation.TPA_Extensions.Flight.OriginLocation"];
                /**
                 * @description Flight type (optional) For the Revalidate Itinerary Functionality. Valid values are A-Air, K-ARUNK, O-Open.
                 * @default A
                 * @example A
                 */
                Type: string;
            }[];
// ...
            SegmentType?: components["schemas"]["OTA_AirLowFareSearchRQ.OriginDestinationInformation.TPA_Extensions.SegmentType"];
// ...
        };
// ...
        "OTA_AirLowFareSearchRQ.TPA_Extensions": {
// ...
            IntelliSellTransaction?: components["schemas"]["OTA_AirLowFareSearchRQ.TPA_Extensions.IntelliSellTransaction"];
// ...
        };
// ...
        "OTA_AirLowFareSearchRQ.TPA_Extensions.IntelliSellTransaction.RequestType": {
// ...
            Name?: string;
        };
// ...
        ScheduleDescType: {
            /** @description Arrival Information. */
            arrival: components["schemas"]["Arrival"];
            /** @description Booking Details. */
            bookingDetails?: components["schemas"]["BookingDetails"];
            /** @description Carrier Information. */
            carrier: components["schemas"]["Carrier"];
            /** @description Departure Information. */
            departure: components["schemas"]["Departure"];
            /**
             * @description DOT Rating.
             * @example A
             */
            dotRating?: string;
            /**
             * @description Informs if ticket is eTicketable.
             * @example true
             */
            eTicketable?: boolean;
            /**
             * @description The elapsedTime at the scheduleDescs level in GIR format. This information allows customers to display travel time for NDC and ATPCO Offers.
             * @example 321
             */
            elapsedTime?: number;
            /**
             * @description Flight Statused.
             * @example 1
             */
            flightStatused?: number;
            /**
             * @description Flight frequency.
             * @example ***W**S
             */
            frequency?: string;
            /**
             * @description Funnel Flight.
             * @example true
             */
            funnel?: boolean;
            /**
             * @description Segment to government Approval.
             * @example true
             */
            governmentApproval?: boolean;
            /** @description Intermediate Point Info. */
            hiddenStops?: components["schemas"]["HiddenStopType"][];
            /**
             * @description Schedule ID.
             * @example 1
             */
            id: number;
            /**
             * @description Message information.
             * @example ABC
             */
            message?: string;
            /**
             * @description Message Type.
             * @example ABC
             */
            messageType?: string;
            /**
             * @description DEI502 On time performance indicator.
             * @example 1
             */
            onTimePerformance?: number;
            /**
             * @description Is Smoking Allowed.
             * @example true
             */
            smokingAllowed?: boolean;
            /**
             * @description Number Of Stops.
             * @example 1
             */
            stopCount?: number;
            /**
             * @description Total Air Miles Flown.
             * @example 10
             */
            totalMilesFlown?: number;
            /**
             * @description Traffic Restriction Code.
             * @example R
             */
            trafficRestriction?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}